Added Role-based Access Control

+++ api/routes/divisions.js
@@ -2,6 +2,9 @@ const express = require("express");
 const router = express.Router();
 const mysqlConnection = require("../../connection");
 const Division = require("../models/division");
+const authenticate = require("../middleware/authenticate");
+const authorize = require("../middleware/authorize");
+const Role = require("../models/role");
 
 router.get("/", (req, res, next) => {
   const query = "SELECT * FROM divisions";
@@ -18,28 +21,33 @@ router.get("/", (req, res, next) => {
   });
 });
 
-router.post("/", (req, res, next) => {
-  const newDivision = new Division({
-    id: null,
-    name: req.body.name,
-    tournamentId: req.body.tournamentId,
-  });
-  const query = "INSERT INTO divisions SET ?";
+router.post(
+  "/",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const newDivision = new Division({
+      id: null,
+      name: req.body.name,
+      tournamentId: req.body.tournamentId,
+    });
+    const query = "INSERT INTO divisions SET ?";
 
-  mysqlConnection.query(query, newDivision, (err, result) => {
-    if (err) {
-      console.log("Inserting new division failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully inserted new division.");
-      res.status(201).json({
-        message: "Successfully inserted new division.",
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, newDivision, (err, result) => {
+      if (err) {
+        console.log("Inserting new division failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully inserted new division.");
+        res.status(201).json({
+          message: "Successfully inserted new division.",
+        });
+      }
+    });
+  }
+);
 
 router.get("/:divisionId", (req, res, next) => {
   const query = "SELECT * FROM divisions WHERE id = " + req.params.divisionId;
@@ -72,40 +80,50 @@ router.get("/:divisionId/teams", (req, res, next) => {
   });
 });
 
-router.patch("/:divisionId", (req, res, next) => {
-  const query = "UPDATE divisions SET ? where id = " + req.params.divisionId;
+router.patch(
+  "/:divisionId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "UPDATE divisions SET ? where id = " + req.params.divisionId;
 
-  mysqlConnection.query(query, req.body, (err, result) => {
-    if (err) {
-      console.log("Patching division failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully patched division.");
-      res.status(200).json({
-        message: "Successfully patched division.",
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, req.body, (err, result) => {
+      if (err) {
+        console.log("Patching division failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully patched division.");
+        res.status(200).json({
+          message: "Successfully patched division.",
+        });
+      }
+    });
+  }
+);
 
-router.delete("/:divisionId", (req, res, next) => {
-  const query = "DELETE FROM divisions WHERE id = " + req.params.divisionId;
+router.delete(
+  "/:divisionId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "DELETE FROM divisions WHERE id = " + req.params.divisionId;
 
-  mysqlConnection.query(query, (err, rows, fields) => {
-    if (err) {
-      console.log("Deleting division failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully deleted division.");
-      res.status(200).json({
-        message: "Successfully deleted division.",
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, (err, rows, fields) => {
+      if (err) {
+        console.log("Deleting division failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully deleted division.");
+        res.status(200).json({
+          message: "Successfully deleted division.",
+        });
+      }
+    });
+  }
+);
 
 module.exports = router;
