+++ api/routes/locations.js
@@ -2,6 +2,9 @@ const express = require("express");
 const router = express.Router();
 const mysqlConnection = require("../../connection");
 const Location = require("../models/location");
+const authenticate = require("../middleware/authenticate");
+const authorize = require("../middleware/authorize");
+const Role = require("../models/role");
 
 router.get("/", (req, res, next) => {
   const query = "SELECT * FROM locations";
@@ -9,7 +12,7 @@ router.get("/", (req, res, next) => {
     if (err) {
       console.log("Getting all locations failed.");
       res.status(500).json({
-        error: err
+        error: err,
       });
     } else {
       console.log("Successfully received all locations.");
@@ -18,28 +21,33 @@ router.get("/", (req, res, next) => {
   });
 });
 
-router.post("/", (req, res, next) => {
-  const newLocation = new Location({
-    id: null,
-    name: req.body.name,
-    description: req.body.description
-  });
-  const query = "INSERT INTO locations SET ?";
+router.post(
+  "/",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const newLocation = new Location({
+      id: null,
+      name: req.body.name,
+      description: req.body.description,
+    });
+    const query = "INSERT INTO locations SET ?";
 
-  mysqlConnection.query(query, newLocation, (err, result) => {
-    if (err) {
-      console.log("Inserting new location failed.");
-      res.status(500).json({
-        error: err
-      });
-    } else {
-      console.log("Successfully inserted new location.");
-      res.status(201).json({
-        message: "Successfully inserted new location."
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, newLocation, (err, result) => {
+      if (err) {
+        console.log("Inserting new location failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully inserted new location.");
+        res.status(201).json({
+          message: "Successfully inserted new location.",
+        });
+      }
+    });
+  }
+);
 
 router.get("/:locationId", (req, res, next) => {
   const query = "SELECT * FROM locations WHERE id = " + req.params.locationId;
@@ -47,7 +55,7 @@ router.get("/:locationId", (req, res, next) => {
     if (err) {
       console.log("Getting location failed.");
       res.status(500).json({
-        error: err
+        error: err,
       });
     } else {
       console.log("Successfully received location.");
@@ -56,40 +64,50 @@ router.get("/:locationId", (req, res, next) => {
   });
 });
 
-router.patch("/:locationId", (req, res, next) => {
-  const query = "UPDATE locations SET ? where id = " + req.params.locationId;
+router.patch(
+  "/:locationId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "UPDATE locations SET ? where id = " + req.params.locationId;
 
-  mysqlConnection.query(query, req.body, (err, result) => {
-    if (err) {
-      console.log("Patching location failed.");
-      res.status(500).json({
-        error: err
-      });
-    } else {
-      console.log("Successfully patched location.");
-      res.status(200).json({
-        message: "Successfully patched location."
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, req.body, (err, result) => {
+      if (err) {
+        console.log("Patching location failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully patched location.");
+        res.status(200).json({
+          message: "Successfully patched location.",
+        });
+      }
+    });
+  }
+);
 
-router.delete("/:locationId", (req, res, next) => {
-  const query = "DELETE FROM locations WHERE id = " + req.params.locationId;
+router.delete(
+  "/:locationId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "DELETE FROM locations WHERE id = " + req.params.locationId;
 
-  mysqlConnection.query(query, (err, rows, fields) => {
-    if (err) {
-      console.log("Deleting location failed.");
-      res.status(500).json({
-        error: err
-      });
-    } else {
-      console.log("Successfully deleted location.");
-      res.status(200).json({
-        message: "Successfully deleted location."
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, (err, rows, fields) => {
+      if (err) {
+        console.log("Deleting location failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully deleted location.");
+        res.status(200).json({
+          message: "Successfully deleted location.",
+        });
+      }
+    });
+  }
+);
 
 module.exports = router;
