+++ api/routes/players.js
@@ -2,6 +2,9 @@ const express = require("express");
 const router = express.Router();
 const mysqlConnection = require("../../connection");
 const Player = require("../models/player");
+const authenticate = require("../middleware/authenticate");
+const authorize = require("../middleware/authorize");
+const Role = require("../models/role");
 
 router.get("/", (req, res, next) => {
   const query = "SELECT * FROM players";
@@ -9,7 +12,7 @@ router.get("/", (req, res, next) => {
     if (err) {
       console.log("Getting all players failed.");
       res.status(500).json({
-        error: err
+        error: err,
       });
     } else {
       console.log("Successfully received all players.");
@@ -18,28 +21,33 @@ router.get("/", (req, res, next) => {
   });
 });
 
-router.post("/", (req, res, next) => {
-  const newPlayer = new Player({
-    id: null,
-    name: req.body.name,
-    teamId: req.body.teamId
-  });
-  const query = "INSERT INTO players SET ?";
+router.post(
+  "/",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const newPlayer = new Player({
+      id: null,
+      name: req.body.name,
+      teamId: req.body.teamId,
+    });
+    const query = "INSERT INTO players SET ?";
 
-  mysqlConnection.query(query, newPlayer, (err, result) => {
-    if (err) {
-      console.log("Inserting new player failed.");
-      res.status(500).json({
-        error: err
-      });
-    } else {
-      console.log("Successfully inserted new player.");
-      res.status(201).json({
-        message: "Successfully inserted new player."
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, newPlayer, (err, result) => {
+      if (err) {
+        console.log("Inserting new player failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully inserted new player.");
+        res.status(201).json({
+          message: "Successfully inserted new player.",
+        });
+      }
+    });
+  }
+);
 
 router.get("/:playerId", (req, res, next) => {
   const query = "SELECT * FROM players WHERE id = " + req.params.playerId;
@@ -47,7 +55,7 @@ router.get("/:playerId", (req, res, next) => {
     if (err) {
       console.log("Getting player failed.");
       res.status(500).json({
-        error: err
+        error: err,
       });
     } else {
       console.log("Successfully received player.");
@@ -56,40 +64,50 @@ router.get("/:playerId", (req, res, next) => {
   });
 });
 
-router.patch("/:playerId", (req, res, next) => {
-  const query = "UPDATE players SET ? where id = " + req.params.playerId;
+router.patch(
+  "/:playerId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "UPDATE players SET ? where id = " + req.params.playerId;
 
-  mysqlConnection.query(query, req.body, (err, result) => {
-    if (err) {
-      console.log("Patching player failed.");
-      res.status(500).json({
-        error: err
-      });
-    } else {
-      console.log("Successfully patched player.");
-      res.status(200).json({
-        message: "Successfully patched player."
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, req.body, (err, result) => {
+      if (err) {
+        console.log("Patching player failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully patched player.");
+        res.status(200).json({
+          message: "Successfully patched player.",
+        });
+      }
+    });
+  }
+);
 
-router.delete("/:playerId", (req, res, next) => {
-  const query = "DELETE FROM players WHERE id = " + req.params.playerId;
+router.delete(
+  "/:playerId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "DELETE FROM players WHERE id = " + req.params.playerId;
 
-  mysqlConnection.query(query, (err, rows, fields) => {
-    if (err) {
-      console.log("Deleting player failed.");
-      res.status(500).json({
-        error: err
-      });
-    } else {
-      console.log("Successfully deleted player.");
-      res.status(200).json({
-        message: "Successfully deleted player."
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, (err, rows, fields) => {
+      if (err) {
+        console.log("Deleting player failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully deleted player.");
+        res.status(200).json({
+          message: "Successfully deleted player.",
+        });
+      }
+    });
+  }
+);
 
 module.exports = router;
