+++ api/routes/teams.js
@@ -2,6 +2,9 @@ const express = require("express");
 const router = express.Router();
 const mysqlConnection = require("../../connection");
 const Team = require("../models/team");
+const authenticate = require("../middleware/authenticate");
+const authorize = require("../middleware/authorize");
+const Role = require("../models/role");
 
 router.get("/", (req, res, next) => {
   const query = "SELECT * FROM teams";
@@ -18,28 +21,33 @@ router.get("/", (req, res, next) => {
   });
 });
 
-router.post("/", (req, res, next) => {
-  const newTeam = new Team({
-    id: null,
-    name: req.body.name,
-    divisionId: req.body.divisionId,
-  });
-  const query = "INSERT INTO teams SET ?";
+router.post(
+  "/",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const newTeam = new Team({
+      id: null,
+      name: req.body.name,
+      divisionId: req.body.divisionId,
+    });
+    const query = "INSERT INTO teams SET ?";
 
-  mysqlConnection.query(query, newTeam, (err, result) => {
-    if (err) {
-      console.log("Inserting new team failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully inserted new team.");
-      res.status(201).json({
-        message: "Successfully inserted new team.",
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, newTeam, (err, result) => {
+      if (err) {
+        console.log("Inserting new team failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully inserted new team.");
+        res.status(201).json({
+          message: "Successfully inserted new team.",
+        });
+      }
+    });
+  }
+);
 
 router.get("/:teamId", (req, res, next) => {
   const query = "SELECT * FROM teams WHERE id = " + req.params.teamId;
@@ -71,40 +79,50 @@ router.get("/:teamId/players", (req, res, next) => {
   });
 });
 
-router.patch("/:teamId", (req, res, next) => {
-  const query = "UPDATE teams SET ? where id = " + req.params.teamId;
+router.patch(
+  "/:teamId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "UPDATE teams SET ? where id = " + req.params.teamId;
 
-  mysqlConnection.query(query, req.body, (err, result) => {
-    if (err) {
-      console.log("Patching team failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully patched team.");
-      res.status(200).json({
-        message: "Successfully patched team.",
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, req.body, (err, result) => {
+      if (err) {
+        console.log("Patching team failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully patched team.");
+        res.status(200).json({
+          message: "Successfully patched team.",
+        });
+      }
+    });
+  }
+);
 
-router.delete("/:teamId", (req, res, next) => {
-  const query = "DELETE FROM teams WHERE id = " + req.params.teamId;
+router.delete(
+  "/:teamId",
+  authenticate,
+  authorize([Role.Admin, Role.Owner]),
+  (req, res, next) => {
+    const query = "DELETE FROM teams WHERE id = " + req.params.teamId;
 
-  mysqlConnection.query(query, (err, rows, fields) => {
-    if (err) {
-      console.log("Deleting team failed.");
-      res.status(500).json({
-        error: err,
-      });
-    } else {
-      console.log("Successfully deleted team.");
-      res.status(200).json({
-        message: "Successfully deleted tournament.",
-      });
-    }
-  });
-});
+    mysqlConnection.query(query, (err, rows, fields) => {
+      if (err) {
+        console.log("Deleting team failed.");
+        res.status(500).json({
+          error: err,
+        });
+      } else {
+        console.log("Successfully deleted team.");
+        res.status(200).json({
+          message: "Successfully deleted tournament.",
+        });
+      }
+    });
+  }
+);
 
 module.exports = router;
